[Refactor][Workloads] Merge engram_fwd/bwd/decode workloads into engram.py
Engram is not an attention op — consolidate the three small workload
files (99 lines total) into a single workloads/ops/engram.py.
